backend/requirements.txt
Django==2.2.3
django-templated-mail==1.1.1
djangorestframework==3.10.2
djangorestframework-jwt==1.8.0
django-redis==4.10.0
djoser==2.0.1
django-templated-mail==1.1.1
mysqlclient==1.4.2.post1
npm==0.1.1
optional-django==0.1.0
pytz==2019.1
sqlparse==0.3.0
gunicorn==19.9.0
djangorestframework-simplejwt==4.3.0
PyJWT==1.7.1
django-cors-headers==3.0.2
django-allauth
bcrypt==3.1.7
facepy==1.0.10
google-auth==1.6.3
Pillow==6.1.0
